rc4 and refactoring
